refactor(api): build calendars metadata in-process, dropping the /calendars HTTP self-call (#659)
RegionalDataHandler, EventsParams, and CalendarParams previously obtained the
calendars metadata index by issuing an HTTP request to the API's own
/calendars endpoint and re-parsing the response via MetadataCalendars::
fromObject(). That self-call is fragile (it produced the two failures fixed in
 #658 — wildcard host and self-throttling), adds latency, consumes a worker,
and runs before the middleware pipeline so its failures bypass
ErrorHandlingMiddleware.

Extract the index-building logic out of MetadataHandler into a reusable
CalendarMetadataProvider service. MetadataHandler (GET /calendars) and all
three former self-call sites now build the index in-process from local source
data — a single source of truth.

The built object is, by construction, identical to what consumers previously
received over the network: the build path populates every field
(*_keys, diocesan_groups, per-nation dioceses lists) that jsonSerialize()
emits and fromObject() reads back, so the JSON round-trip is idempotent.
CalendarMetadataProviderTest asserts this byte-for-byte, guaranteeing
"use the built object directly" is indistinguishable from "fetch it over HTTP".

Health.php's /calendars call is intentionally retained — it deliberately
exercises the live endpoint as a health check. The wildcard-host normalization
and loopback rate-limit exemption from #658 remain valid for any other
self-calls but are no longer required by these paths.

Tests that previously routed the self-call at a file:// fixture now build from
the bundled sourcedata (Router::$apiFilePath), matching production.
